# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "articleManger",
        "contact": {
            "name": "mylxsw",
            "email": "mylxsw@aicode.cc"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "開發環境"
        }
    ],
    "paths": {
...
        "/article/create": {
            "post": {
                "tags": [
                    "Article"
                ],
                "summary": "創建新分類",
                "operationId": "6830b019af457cf98618e1b6fc8c8f09",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "文章分類",
                                        "type": "string"
                                    },
                                    "article": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "title": {
                                                    "description": "文章標題",
                                                    "type": "string"
                                                },
                                                "content": {
                                                    "description": "文章內容",
                                                    "type": "string"
                                                },
                                                "language": {
                                                    "description": "文章語言",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "創建成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/article/{id}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Get article by ID",
                "operationId": "392a3f313ef2fd63cc716612d1fe93f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Article ID",
                                            "type": "integer"
                                        },
                                        "title": {
                                            "description": "Article title",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/article/update": {
            "put": {
                "tags": [
                    "Article"
                ],
                "summary": "更新分類",
                "operationId": "25998eeeef0aca77269f90640e5ecae8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Article ID",
                                        "type": "integer"
                                    },
                                    "title": {
                                        "description": "Article title",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新完成",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "查無此資料"
                    }
                }
            }
        },
        "/article/delete/{id}": {
            "delete": {
                "tags": [
                    "Article"
                ],
                "summary": "刪除分類",
                "operationId": "e9a537d5c720d1e9da0d9565f760c8ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刪除完成",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}